add: tensorflow-keras-based image classifier.
This classifier uses the keras `model.fit` api and predicts on multiple images in the same process. This makes it *much* faster than the current `TensorFlowInceptionV3Extractor`, which calls a separate python script on each image. The class added by this commit only has to do the TensorFlow setup things once, so only the first predict is impacted. All subsequent predictions are fast.